Implement update_coded_lines
Used SqlAlchemy's ORM mode to accomplish this.
Fetches all existing coded lines for the document and coder, and
then compares the set of existing coded line data with new coded line data.
When existing are absent from new, marks objects for deletion.
When new are absent from existing, creates a new CodedLine and adds it to the session.