Remove sql uniqueness constraint on CodedLine
Previously, there was a sql uniqueness constraint on CodedLine
of (line, coder, code). This can't work; the same coder could use
the same code on the same line of different documents. The actual
constraint ought to be uniqueness of (line, coder, code) when
line.location.document is the same, but this would be difficult
to express in sql.